+++ app/src/main/java/app/passwordstore/ui/onboarding/fragments/WelcomeFragment.kt
@@ -9,7 +9,6 @@ import android.os.Bundle
 import android.view.View
 import androidx.annotation.Keep
 import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
 import app.passwordstore.R
 import app.passwordstore.databinding.FragmentWelcomeBinding
@@ -17,6 +16,7 @@ import app.passwordstore.ui.settings.SettingsActivity
 import app.passwordstore.util.extensions.launchActivity
 import app.passwordstore.util.extensions.performTransactionWithBackStack
 import app.passwordstore.util.extensions.viewBinding
+import app.passwordstore.util.extensions.windowInsetsLambda
 
 @Keep
 class WelcomeFragment : Fragment(R.layout.fragment_welcome) {
@@ -26,11 +26,7 @@ class WelcomeFragment : Fragment(R.layout.fragment_welcome) {
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
 
-    ViewCompat.setOnApplyWindowInsetsListener(view) { v, insets ->
-      val statusBarHeight = insets.getInsets(WindowInsetsCompat.Type.statusBars()).top
-      v.setPadding(0, statusBarHeight, 0, 0)
-      insets
-    }
+    ViewCompat.setOnApplyWindowInsetsListener(view, windowInsetsLambda)
 
     with(binding) {
       letsGo.setOnClickListener {